+++ tricolour/apps/tricolour/strat_executor.py
@@ -1,9 +1,5 @@
 # -*- coding: utf-8 -*-
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
 import dask.array as da
 
 from tricolour.dask_wrappers import (sum_threshold_flagger,