+++ tricolour/tests/test_stokes.py
@@ -1,9 +1,5 @@
 # -*- coding: utf-8 -*-
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
 import numpy as np
 import pytest
 
@@ -21,7 +17,7 @@ from tricolour.stokes import (stokes_corr_map,
 def test_unpolarised_intensity(stokes):
     # Set up our stokes parameters in an interesting order
     stokes = list(map(STOKES_TYPES.__getitem__, stokes))
-    vis = np.asarray([[[1+1j, 2+2j, 3+3j, 4+4j]]], np.complex128)
+    vis = np.asarray([[[1 + 1j, 2 + 2j, 3 + 3j, 4 + 4j]]], np.complex128)
 
     stokes_map = stokes_corr_map(stokes)
 
@@ -30,16 +26,16 @@ def test_unpolarised_intensity(stokes):
     unpol = 0
 
     for c1, c2, a, s1, s2 in stokes_unpol:
-        v = a*(s1*vis[0, 0, c1] + s2*vis[0, 0, c2])
-        unpol += v.real  # imaginary contains only noise
+        v = a * (s1 * vis[0, 0, c1] + s2 * vis[0, 0, c2])
+        unpol += np.abs(v)
 
     # Polarised stokes mappings
     stokes_pol = tuple(v for k, v in stokes_map.items() if k != 'I')
     pol = 0
 
     for c1, c2, a, s1, s2 in stokes_pol:
-        v = a*(s1*vis[0, 0, c1] + s2*vis[0, 0, c2])
-        pol += v.real**2  # imaginary contains only noise
+        v = a * (s1 * vis[0, 0, c1] + s2 * vis[0, 0, c2])
+        pol += np.abs(v)**2
 
     upi = unpol - np.sqrt(pol)
     val = unpolarised_intensity(vis, stokes_unpol, stokes_pol)
@@ -54,7 +50,7 @@ def test_unpolarised_intensity(stokes):
 def test_polarised_intensity(stokes):
     # Set up our stokes parameters in an interesting order
     stokes = list(map(STOKES_TYPES.__getitem__, stokes))
-    vis = np.asarray([[[1+1j, 2+2j, 3+3j, 4+4j]]], np.complex128)
+    vis = np.asarray([[[1 + 1j, 2 + 2j, 3 + 3j, 4 + 4j]]], np.complex128)
 
     stokes_map = stokes_corr_map(stokes)
 
@@ -63,8 +59,8 @@ def test_polarised_intensity(stokes):
     pol = 0
 
     for c1, c2, a, s1, s2 in stokes_pol:
-        v = a*(s1*vis[0, 0, c1] + s2*vis[0, 0, c2])
-        pol += v.real**2  # imaginary contains only noise
+        v = a * (s1 * vis[0, 0, c1] + s2 * vis[0, 0, c2])
+        pol += np.abs(v)**2  # imaginary contains only noise
 
     pi = np.sqrt(pol)
     val = polarised_intensity(vis, stokes_pol)