+++ tricolour/window_statistics.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 from collections import defaultdict
 from functools import partial
 
@@ -20,35 +22,46 @@ def _window_stats(flag_window, ubls, chan_freqs,
 
     stats = WindowStatistics(nchanbins)
 
+    # per antenna
     for ai, a in enumerate(antenna_names):
-        # per antenna
         sel = np.logical_or(ubls[:, 1] == ai, ubls[:, 2] == ai)
-        cnt = np.sum(flag_window[sel, :, :, :])
+        cnt = np.sum(flag_window[sel, :, :, :], dtype=np.uint64)
         sz = flag_window[sel, :, :, :].size
         stats._counts_per_ant[a] += cnt
         stats._size_per_ant[a] += sz
 
-        # per scan and field
-        cnt = np.sum(flag_window)
-        sz = flag_window.size
-        stats._counts_per_field[field_name] += cnt
-        stats._size_per_field[field_name] += sz
-        stats._counts_per_scan[scan_no] += cnt
-        stats._size_per_scan[scan_no] += sz
+    # per baseline
+    for bi, b in enumerate(np.unique(ubls[:, 0])):
+        sel = ubls[:, 0] == b
+        sela1 = antenna_names[ubls[sel, 1][0]]
+        sela2 = antenna_names[ubls[sel, 2][0]]
+        blname = "{0:s}&{1:s}".format(sela1, sela2)
+        cnt = np.sum(flag_window[sel, :, :, :], dtype=np.uint64)
+        sz = flag_window[sel, :, :, :].size
+        stats._counts_per_bl[blname] += cnt
+        stats._size_per_bl[blname] += sz
 
-        # binned per channel
-        bins_edges = np.linspace(np.min(chan_freqs), np.max(chan_freqs),
-                                 nchanbins)
-        bins = np.zeros(nchanbins, dtype=np.uint32)
+    # per scan and field
+    cnt = np.sum(flag_window, dtype=np.uint64)
+    sz = flag_window.size
+    stats._counts_per_field[field_name] += cnt
+    stats._size_per_field[field_name] += sz
+    stats._counts_per_scan[scan_no] += cnt
+    stats._size_per_scan[scan_no] += sz
 
-        for ch_i, ch in enumerate(bins_edges[:-1]):
-            sel = np.logical_and(chan_freqs >= bins_edges[ch_i],
-                                 chan_freqs < bins_edges[ch_i + 1])
-            bins[ch_i] = np.sum(flag_window[:, :, :, sel])
+    # binned per channel
+    bins_edges = np.linspace(np.min(chan_freqs), np.max(chan_freqs),
+                             nchanbins)
+    bins = np.zeros(nchanbins, dtype=np.uint32)
 
-        stats._counts_per_ddid[ddid] += bins
-        stats._bins_per_ddid[ddid] = bins_edges  # frequency labels
-        stats._size_per_ddid[ddid] += flag_window.size
+    for ch_i, ch in enumerate(bins_edges[:-1]):
+        sel = np.logical_and(chan_freqs >= bins_edges[ch_i],
+                             chan_freqs < bins_edges[ch_i + 1])
+        bins[ch_i] = np.sum(flag_window[:, :, :, sel], dtype=np.uint64)
+
+    stats._counts_per_ddid[ddid] += bins
+    stats._bins_per_ddid[ddid] = bins_edges  # frequency labels
+    stats._size_per_ddid[ddid] += flag_window.size
 
     return stats
 
@@ -163,11 +176,13 @@ class WindowStatistics(object):
         self._counts_per_ant = defaultdict(lambda: 0)
         self._counts_per_field = defaultdict(lambda: 0)
         self._counts_per_scan = defaultdict(lambda: 0)
+        self._counts_per_bl = defaultdict(lambda: 0)
         self._size_per_ant = defaultdict(lambda: 0)
         self._size_per_field = defaultdict(lambda: 0)
         self._size_per_scan = defaultdict(lambda: 0)
+        self._size_per_bl = defaultdict(lambda: 0)
 
-        bin_factory = partial(np.zeros, nchanbins, dtype=np.uint32)
+        bin_factory = partial(np.zeros, nchanbins, dtype=np.uint64)
         self._counts_per_ddid = defaultdict(bin_factory)
         self._bins_per_ddid = defaultdict(lambda: 0)
         self._size_per_ddid = defaultdict(lambda: 0)
@@ -186,6 +201,9 @@ class WindowStatistics(object):
         for d, count in other._counts_per_ddid.items():
             self._counts_per_ddid[d] += count
 
+        for b, count in other._counts_per_bl.items():
+            self._counts_per_bl[b] += count
+
         # Sizes
         for a, size in other._size_per_ant.items():
             self._size_per_ant[a] += size
@@ -198,6 +216,9 @@ class WindowStatistics(object):
 
         for s, size in other._size_per_ddid.items():
             self._size_per_ddid[s] += size
+
+        for b, size in other._size_per_bl.items():
+            self._size_per_bl[b] += size
 
         # ddid
         for d, bins in other._bins_per_ddid.items():
@@ -225,36 +246,45 @@ def summarise_stats(final, original):
 
     l.append("Per antenna:")
     for a in final._counts_per_ant:
-        l.append("\t {0:s}: {1:.3f}%, original {2:.3f}%".format(a,
-                 final._counts_per_ant[a] * 100.0 /
-                 final._size_per_ant[a],
-                 original._counts_per_ant[a] * 100.0 /
-                 original._size_per_ant[a]))
+        l.append("\t {0:s}: {1:.3f}%, original {2:.3f}%".format(
+            a, final._counts_per_ant[a] * 100.0 /
+            final._size_per_ant[a],
+            original._counts_per_ant[a] * 100.0 /
+            original._size_per_ant[a]))
+
     l.append("Per scan:")
     for s in final._counts_per_scan:
-        l.append("\t {0:d}: {1:.3f}%, original {2:.3f}%".format(s,
-                 final._counts_per_scan[s] * 100.0 /
-                 final._size_per_scan[s],
-                 original._counts_per_scan[s] * 100.0 /
-                 original._size_per_scan[s]))
+        l.append("\t {0:d}: {1:.3f}%, original {2:.3f}%".format(
+            s, final._counts_per_scan[s] * 100.0 /
+            final._size_per_scan[s],
+            original._counts_per_scan[s] * 100.0 /
+            original._size_per_scan[s]))
 
     l.append("Per field:")
     for f in final._counts_per_field:
-        l.append("\t {0:s}: {1:.3f}%, original {2:.3f}%".format(f,
-                 final._counts_per_field[f] * 100.0 /
-                 final._size_per_field[f],
-                 original._counts_per_field[f] * 100.0 /
-                 original._size_per_field[f]))
+        l.append("\t {0:s}: {1:.3f}%, original {2:.3f}%".format(
+            f, final._counts_per_field[f] * 100.0 /
+            final._size_per_field[f],
+            original._counts_per_field[f] * 100.0 /
+            original._size_per_field[f]))
+
+    l.append("Per baseline:")
+    for b in final._counts_per_bl:
+        l.append("\t {0:s}: {1:.3f}%, original {2:.3f}%".format(
+            b, final._counts_per_bl[b] * 100.0 /
+            final._size_per_bl[b],
+            original._counts_per_bl[b] * 100.0 /
+            original._size_per_bl[b]))
 
     l.append("Per data descriptor id:")
 
     for d in final._counts_per_ddid:
         ratios = final._counts_per_ddid[d] * 100.0 / final._size_per_ddid[d]
-        ratio_str = '\t'.join(["{0:.2f}".format(r) for r in ratios])
+        ratio_str = '\t'.join(["{0:<7.2f}".format(r) for r in ratios])
         l.append("\t {0:d}: {1:s}%".format(d, ratio_str))
 
         ddid_freqs = final._bins_per_ddid[d] / 1e6
-        ddid_freqs_str = '\t'.join(["{0:.1f}".format(f) for f in ddid_freqs])
+        ddid_freqs_str = '\t'.join(["{0:<7.1f}".format(f) for f in ddid_freqs])
         l.append("\t    {0:s} MHz".format(ddid_freqs_str))
 
     l.append("********************************")