+++ HISTORY.rst
@@ -2,9 +2,24 @@
 History
 =======
 
-0.1.3 (YYYY-MM-DD)
+0.1.7 (2020-01-06)
 ------------------
 
+* Change license to BSD3 (:pr:`64`)
+
+0.1.4 (2019-07-23)
+------------------
+
+* Upgrade to dask-ms (:pr:`59`)
+
+0.1.3 (2019-07-22)
+------------------
+
+* Use pytest-flake8 when running test cases (:pr:`56`)
+* Fix polarisation flagging for uncalibrated data (:pr:`55`)
+* Add ability to flag on total power (:pr:`55`)
+* Baseline statistics (:pr:`55`)
+* 4K UHF mask (:pr:`55`)
 * Add Pull Request Template (:pr:`53`)